chore: rename mocks file
