+++ plugins/backstage-plugin-coder/src/components/CoderWorkspacesCard/WorkspacesList.test.tsx
@@ -3,7 +3,7 @@ import { type WorkspacesListProps, WorkspacesList } from './WorkspacesList';
 import { renderInCoderEnvironment } from '../../testHelpers/setup';
 import { CardContext, WorkspacesCardContext, WorkspacesQuery } from './Root';
 import { mockCoderWorkspacesConfig } from '../../testHelpers/mockBackstageData';
-import { mockWorkspaceWithMatch } from '../../testHelpers/mockCoderAppData';
+import { mockWorkspaceWithMatch } from '../../testHelpers/mockCoderPluginData';
 import type { Workspace } from '../../api/vendoredSdk';
 import { screen } from '@testing-library/react';
