+++ plugins/backstage-plugin-coder/src/testHelpers/mockCoderPluginData.ts
